stay with floats for qReal for now. Should fix painting issues on win.
[git-p4: depot-paths = "//depot/qt/main/": change = 163632]

+++ src/core/global/qglobal.h
@@ -1819,7 +1819,7 @@ inline int qIntCast(float f) { return int(f); }
 #include <qfixedpoint.h>
 typedef QFixedPoint qReal;
 #else
-typedef double qReal;
+typedef float qReal;
 #endif
 
 #endif /* __cplusplus */
